+++ src/main/java/gov/nasa/cumulus/metadata/aggregator/constant/MENDsIsoXMLSpatialTypeConstant.java
@@ -1,0 +1,11 @@
+package gov.nasa.cumulus.metadata.aggregator.constant;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class MENDsIsoXMLSpatialTypeConstant {
+    public static final String FOOTPRINT = "footprint";
+    public static final String BBOX = "bbox";
+    public static final String ORBIT = "orbit";
+    public static final List<String> isoXMLSpatialTypeList = Arrays.asList(FOOTPRINT, BBOX, ORBIT);
+}